+++ client/src/components/Auth/Auth.js
@@ -1,11 +1,11 @@
 import React from "react";
-import Login from "./Login";
-import Signup from "./Signup";
+import Login from "./Login/Login";
+import Signup from "./Signup/Signup";
 import classes from "./Auth.module.css";
 
 class Auth extends React.Component {
   state = {
-    isLogginIn: false,
+    switchAuthType: false,
     authComponentState: {},
   };
 
@@ -16,35 +16,43 @@ class Auth extends React.Component {
     fontFamily: "Helvetica",
   };
 
+  //////////////////////
+  // props received
+  // isLoggedIn
+
   render() {
     return (
       <div className={classes.Container}>
         {/* Form Logo Goes Here */}
         <div className={classes.AdSide}>
-          <div>
-            <p className={classes.Para}>DigiPharm</p>
-            {/* <img src="./health.png" alt="" /> */}
-          </div>
+          <p className={classes.Para}>DigiPharm</p>
+          {/* <img src="./health.png" alt="" /> */}
         </div>
+
         {/* Form Signin - Signup Goes Here */}
         <div className={classes.AuthSide}>
           <div className={classes.FormToggle}>
+            <div className={classes.Spacer}></div>
             <button
-              style={!this.state.isLogginIn ? this.activeStyle : {}}
-              onClick={() => {
-                this.setState({ isLogginIn: false });
-              }}
+              className={classes.FormToggle__Button}
+              style={!this.state.switchAuthType ? this.activeStyle : {}}
+              onClick={() => this.setState({ switchAuthType: false })}
             >
               Signup
             </button>
             <button
-              style={this.state.isLogginIn ? this.activeStyle : {}}
-              onClick={() => this.setState({ isLogginIn: true })}
+              className={classes.FormToggle__Button}
+              style={this.state.switchAuthType ? this.activeStyle : {}}
+              onClick={() => this.setState({ switchAuthType: true })}
             >
               Login
             </button>
           </div>
-          {this.state.isLogginIn ? <Login /> : <Signup />}
+          {this.state.switchAuthType ? (
+            <Login logInComplete={this.isLoggedIn} />
+          ) : (
+            <Signup />
+          )}
         </div>
       </div>
     );
