+++ client/src/components/Auth/Auth.module.css
@@ -1,69 +1,89 @@
-* {
-  text-align: center;
-}
-
 .Container {
   display: flex;
   flex-direction: row;
-  justify-items: center;
-  justify-content: center;
-  width: 70%;
-  margin: 5% 12.5% 0 12.5%;
-  height: 60%;
+  width: 80%;
+  margin: 5% auto auto auto;
+  height: 80%;
   box-shadow: 2px 2px 4px grey;
   padding: 2% 2.5%;
-}
-
-.Radio,
-.Radio:hover {
-  border: 1px solid aqua;
-  cursor: pointer;
-  margin-top: 22px;
+  box-sizing: border-box;
 }
 
 .AdSide {
   background-color: aquamarine;
-  width: 50%;
-  min-height: 500px;
-}
-
-.AuthSide {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-}
-
-.FormToggle {
-  margin-bottom: 10px;
-  position: relative;
-  text-align: right;
-  margin-right: 34px;
-}
-
-.FormToggle > button {
-  height: 38px;
-  font-size: 16px;
-  border: none;
-  cursor: pointer;
-  width: 100px;
-  border: 1px solid aqua;
+  text-align: center;
+  flex: 1;
 }
 
 .Para {
   font-size: 28px;
   font-weight: 900;
   color: lightcoral;
+  align-self: center;
+  margin: 0 auto;
   font-family: Helvetica, sans-serif;
+}
+
+.AuthSide {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 0;
+  background-color: rebeccapurple;
+  flex: 2;
+}
+
+.Spacer {
+  flex: 1;
+}
+
+.FormToggle {
+  display: flex;
+  margin-bottom: 15px;
+  width: 100%;
+  text-align: right;
+}
+
+.FormToggle__Button {
+  height: 38px;
+  font-size: 16px;
+  border: none;
+  cursor: pointer;
+  width: 100px;
+  border: 1px solid aqua;
+  border-radius: 0px;
+}
+
+.FormToggle__Button_Active {
+  background-color: aqua;
+  color: white;
+  border: none;
+  font-family: Helvetica;
+}
+
+/* .Radio,
+.Radio:hover {
+  border: 1px solid aqua;
+  cursor: pointer;
+  margin-top: 22px;
+} */
+
+.FormGroup {
+  display: flex;
+  flex-direction: column;
+  width: 90%;
+  align-items: center;
+  text-align: center;
 }
 
 .InputItem {
   display: block;
   height: 32px;
-  width: 85%;
-  margin: 2% 6% 2% auto;
-  position: relative;
+  width: 100%;
+  margin: 3% auto;
   border-radius: 4px;
   border: none;
+  text-align: center;
   border-bottom: 1px solid aqua;
 }
 
@@ -76,11 +96,9 @@
   border-bottom: 1px solid forestgreen;
 }
 
-.Button {
+.Button__Submit {
   width: 100px;
   height: 48px;
-  margin-top: 25px;
-  margin-bottom: 25px;
   font-size: 18px;
   font-family: Arial, Helvetica, sans-serif;
   font-weight: bolder;
@@ -90,8 +108,27 @@
   color: white;
 }
 
-.Button:hover {
+.Button__Submit:hover {
   background-color: white;
   border: 1px solid aqua;
   color: aqua;
+}
+
+@media (max-width: 768px) {
+  .AdSide {
+    display: none;
+  }
+
+  .Container {
+    padding-top: 25px;
+  }
+
+  .Spacer {
+    display: none;
+  }
+
+  .FormToggle__Button {
+    flex: 1;
+    height: 25px;
+  }
 }
